requirements.txt
pytest-bdd~=5.0.0
py~=1.11.0
colorama~=0.4.6
pytest~=7.4.2
h11~=0.14.0
pip~=23.2.1
attrs~=23.1.0
wheel~=0.41.2
setuptools~=68.2.0
Mako~=1.3.2
MarkupSafe~=2.1.5
pony==0.7.17
trio~=0.22.2
outcome~=1.2.0
sniffio~=1.3.0
sortedcontainers~=2.4.0
cffi~=1.16.0
idna~=3.4
mysql-connector-python~=8.3.0
pluggy~=1.3.0
iniconfig~=2.0.0
certifi~=2023.7.22
selenium==3.141.0
urllib3~=2.0.6
pycparser~=2.21
parse~=1.20.1
six~=1.16.0
glob2~=0.7
packaging~=23.2
wsproto~=1.2.0
PyMySQL~=1.1.0
jsonpickle~=3.0.2
allure-pytest